Lukicheva Polina Ex04

+++ include/task1.h
@@ -10,7 +10,7 @@ struct Person
 };
 
 int comp1(const void* a, const void* b);
-int comp2(const void* a, const void* b);
+double comp2(const void* a, const void* b);
 int comp3(const void* a, const void* b);
 int comp4(const void* a, const void* b);
 int comp5(const void* a, const void* b);
